// data/foodb-models/db.d.ts
export interface referencesAttribute {
  id:number;
  refType?:string;
  text?:string;
  pubmedId?:string;
  link?:string;
  title?:string;
  creatorId?:number;
  updaterId?:number;
  createdAt:Date;
  updatedAt:Date;
  sourceId?:number;
  sourceType?:string;
}
export interface referencesInstance extends Sequelize.Instance<referencesAttribute>, referencesAttribute { }